requirements.txt
customtkinter>=5.2.0
playwright>=1.40.0
requests>=2.31.0
Pillow>=10.0.0
psutil>=5.9.0
